% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quid.R
\docType{package}
\name{quid}
\alias{quid}
\alias{quid-package}
\title{Bayesian Mixed Models for Qualitative Individual Differences}
\description{
Estimate individual effect sizes and evaluate order constraints
}
\author{
\strong{Maintainer}: Lukas Klima \email{lukas.klima@student.uva.nl}

Authors:
\itemize{
  \item Julia Haaf \email{j.m.haaf@uva.nl} [degree supervisor, thesis advisor]
}

}
