% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quids.R
\name{quids}
\alias{quids}
\title{Qualitative Individual Differences (Stroop)}
\usage{
quids(
  id,
  condition,
  rt,
  prior = c(1/6, 1/10),
  expect = list(lower = 1, higher = 2),
  iter = 10000,
  burnin = 1000,
  messages = TRUE
)
}
\arguments{
\item{id}{Numeric or factor. A vector indicating the participant ID.}

\item{condition}{Numeric or factor. A vector indicating the condition.}

\item{rt}{Numeric. A vector containing the response times for each trial.}
}
\value{
This function returns a \code{list} containing the following elements:
\itemize{
 \item theta
 \item BF
}
}
\description{
This quid function can be used to estimate individual effect sizes and to
evaluate order constraints in settings with two conditions, one experimental
and one control.
}
\section{Warning}{

\code{id}, \code{condition} and \code{rt} all must have the same length.
}

