% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quid.R
\docType{package}
\name{quid}
\alias{quid}
\alias{quid-package}
\title{Bayesian Mixed Models for Qualitative Individual Differences}
\description{
Test whether equality and order constraints hold for all individuals
simultaneously by comparing Bayesian mixed models through Bayes factors.
}
\author{
\strong{Maintainer}: Lukas Klima \email{lukas.klima@live.at}

Authors:
\itemize{
  \item Julia Haaf \email{j.m.haaf@uva.nl} [degree supervisor, thesis advisor]
}

}
